fix: The response content cannot be parsed because the Internet Explorer engine is not available
There is an error when trying to run the script: `Error downloading MachineKeys.txt: The response content cannot be parsed because the Internet Explorer engine is not available, or Internet Explorer's first-launch configuration is not complete. Specify the UseBasicParsing parameter and try again`

When adding the `UseBasicParameter` the script works.